+++ webapp/javascript/pages/ContinuousComparisonView.tsx
@@ -11,6 +11,7 @@ import {
   fetchComparisonSide,
   fetchTagValues,
   selectQueries,
+  selectTimelineSides,
 } from '@webapp/redux/reducers/continuous';
 import TimelineChartWrapper from '@webapp/components/TimelineChart/TimelineChartWrapper';
 import SyncTimelines from '@webapp/components/TimelineChart/SyncTimelines';
@@ -45,9 +46,17 @@ function ComparisonApp() {
   const { colorMode } = useColorMode();
   usePopulateLeftRightQuery();
   const comparisonView = useAppSelector(selectComparisonState);
-  const { leftTags, rightTags } = useTags({ leftQuery, rightQuery });
+  const { leftTags, rightTags } = useTags();
   const { leftTimeline, rightTimeline } = useTimelines();
   const sharedQuery = useFlamegraphSharedQuery();
+
+  const timelines = useAppSelector(selectTimelineSides);
+  const isLoading = isLoadingOrReloading([
+    comparisonView.left.type,
+    comparisonView.right.type,
+    timelines.left.type,
+    timelines.right.type,
+  ]);
 
   useEffect(() => {
     if (leftQuery) {
@@ -86,18 +95,12 @@ function ComparisonApp() {
       <PageTitle title={formatTitle('Comparison', leftQuery, rightQuery)} />
       <div className="main-wrapper">
         <Toolbar
-          hideTagsBar
-          onSelectedName={(query) => {
+          onSelectedApp={(query) => {
             dispatch(actions.setQuery(query));
           }}
         />
         <Box>
-          <LoadingOverlay
-            active={isLoadingOrReloading([
-              comparisonView.left.type,
-              comparisonView.right.type,
-            ])}
-          >
+          <LoadingOverlay active={isLoading}>
             <TimelineChartWrapper
               data-testid="timeline-main"
               id="timeline-chart-double"
@@ -145,20 +148,13 @@ function ComparisonApp() {
           data-testid="comparison-container"
         >
           <Box className={styles.comparisonPane}>
-            <LoadingOverlay
-              active={isLoadingOrReloading([comparisonView.left.type])}
-              spinnerPosition="baseline"
-            >
+            <LoadingOverlay active={isLoading} spinnerPosition="baseline">
               <TimelineTitle titleKey="baseline" color={leftColor} />
               <TagsBar
                 query={leftQuery}
                 tags={leftTags}
-                onSetQuery={(q) => {
-                  dispatch(actions.setLeftQuery(q));
-                  if (leftQuery === q) {
-                    dispatch(actions.refresh());
-                  }
-                }}
+                onRefresh={() => dispatch(actions.refresh())}
+                onSetQuery={(q) => dispatch(actions.setLeftQuery(q))}
                 onSelectedLabel={(label, query) => {
                   dispatch(fetchTagValues({ query, label }));
                 }}
@@ -209,20 +205,13 @@ function ComparisonApp() {
           </Box>
 
           <Box className={styles.comparisonPane}>
-            <LoadingOverlay
-              spinnerPosition="baseline"
-              active={isLoadingOrReloading([comparisonView.right.type])}
-            >
+            <LoadingOverlay spinnerPosition="baseline" active={isLoading}>
               <TimelineTitle titleKey="comparison" color={rightColor} />
               <TagsBar
                 query={rightQuery}
                 tags={rightTags}
-                onSetQuery={(q) => {
-                  dispatch(actions.setRightQuery(q));
-                  if (rightQuery === q) {
-                    dispatch(actions.refresh());
-                  }
-                }}
+                onRefresh={() => dispatch(actions.refresh())}
+                onSetQuery={(q) => dispatch(actions.setRightQuery(q))}
                 onSelectedLabel={(label, query) => {
                   dispatch(fetchTagValues({ query, label }));
                 }}