+++ webapp/javascript/pages/ContinuousDiffView.tsx
@@ -47,7 +47,7 @@ function ComparisonDiffApp() {
   const { leftQuery, rightQuery } = useAppSelector(selectQueries);
 
   usePopulateLeftRightQuery();
-  const { leftTags, rightTags } = useTags({ leftQuery, rightQuery });
+  const { leftTags, rightTags } = useTags();
   const { leftTimeline, rightTimeline } = useTimelines();
 
   const timelines = useAppSelector(selectTimelineSides);
@@ -57,6 +57,12 @@ function ComparisonDiffApp() {
 
   const { offset } = useTimeZone();
   const timezone = offset === 0 ? 'utc' : 'browser';
+
+  const isLoading = isLoadingOrReloading([
+    diffView.type,
+    timelines.left.type,
+    timelines.right.type,
+  ]);
 
   useEffect(() => {
     if (rightQuery && leftQuery) {
@@ -102,18 +108,12 @@ function ComparisonDiffApp() {
       <PageTitle title={formatTitle('Diff', leftQuery, rightQuery)} />
       <div className="main-wrapper">
         <Toolbar
-          hideTagsBar
-          onSelectedName={(query) => {
+          onSelectedApp={(query) => {
             dispatch(actions.setQuery(query));
           }}
         />
         <Box>
-          <LoadingOverlay
-            active={isLoadingOrReloading([
-              timelines.left.type,
-              timelines.right.type,
-            ])}
-          >
+          <LoadingOverlay active={isLoading}>
             <TimelineChartWrapper
               data-testid="timeline-main"
               id="timeline-chart-diff"
@@ -156,19 +156,13 @@ function ComparisonDiffApp() {
         </Box>
         <div className="diff-instructions-wrapper">
           <Box className="diff-instructions-wrapper-side">
-            <LoadingOverlay
-              active={isLoadingOrReloading([timelines.left.type])}
-            >
+            <LoadingOverlay active={isLoading}>
               <TimelineTitle titleKey="baseline" color={leftColor} />
               <TagsBar
                 query={leftQuery}
                 tags={leftTags}
-                onSetQuery={(q) => {
-                  dispatch(actions.setLeftQuery(q));
-                  if (leftQuery === q) {
-                    dispatch(actions.refresh());
-                  }
-                }}
+                onRefresh={() => dispatch(actions.refresh())}
+                onSetQuery={(q) => dispatch(actions.setLeftQuery(q))}
                 onSelectedLabel={(label, query) => {
                   dispatch(fetchTagValues({ query, label }));
                 }}
@@ -196,19 +190,13 @@ function ComparisonDiffApp() {
             </LoadingOverlay>
           </Box>
           <Box className="diff-instructions-wrapper-side">
-            <LoadingOverlay
-              active={isLoadingOrReloading([timelines.right.type])}
-            >
+            <LoadingOverlay active={isLoading}>
               <TimelineTitle titleKey="comparison" color={rightColor} />
               <TagsBar
                 query={rightQuery}
                 tags={rightTags}
-                onSetQuery={(q) => {
-                  dispatch(actions.setRightQuery(q));
-                  if (rightQuery === q) {
-                    dispatch(actions.refresh());
-                  }
-                }}
+                onRefresh={() => dispatch(actions.refresh())}
+                onSetQuery={(q) => dispatch(actions.setRightQuery(q))}
                 onSelectedLabel={(label, query) => {
                   dispatch(fetchTagValues({ query, label }));
                 }}
@@ -237,10 +225,7 @@ function ComparisonDiffApp() {
           </Box>
         </div>
         <Box>
-          <LoadingOverlay
-            active={isLoadingOrReloading([diffView.type])}
-            spinnerPosition="baseline"
-          >
+          <LoadingOverlay active={isLoading} spinnerPosition="baseline">
             <TimelineTitle titleKey="diff" />
             <FlamegraphRenderer
               showCredit={false}