+++ webapp/javascript/pages/ExemplarsSingleView.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react';
 import useColorMode from '@webapp/hooks/colorMode.hook';
 import useTimeZone from '@webapp/hooks/timeZone.hook';
 import { useAppSelector, useAppDispatch } from '@webapp/redux/hooks';
-import { selectQueries } from '@webapp/redux/reducers/continuous';
+import { selectQueries, setQuery } from '@webapp/redux/reducers/continuous';
 import {
   fetchExemplarsSingleView,
   fetchSelectionProfile,
@@ -15,9 +15,11 @@ import PageTitle from '@webapp/components/PageTitle';
 import { Heatmap } from '@webapp/components/Heatmap';
 import ExportData from '@webapp/components/ExportData';
 import LoadingSpinner from '@webapp/ui/LoadingSpinner';
+import StatusMessage from '@webapp/ui/StatusMessage';
 import { DEFAULT_HEATMAP_PARAMS } from '@webapp/components/Heatmap/constants';
 import { FlamegraphRenderer } from '@pyroscope/flamegraph/src/FlamegraphRenderer';
 import { formatTitle } from './formatTitle';
+import heatmapSelectionGif from './heatmapSelection.gif';
 
 import styles from './ExemplarsSingleView.module.scss';
 
@@ -37,6 +39,7 @@ function ExemplarsSingleView() {
           query,
           from,
           until,
+          shouldFetchProfile: !!exemplarsSingleView.profile,
           ...DEFAULT_HEATMAP_PARAMS,
         })
       );
@@ -71,20 +74,22 @@ function ExemplarsSingleView() {
       case 'loaded':
       case 'reloading': {
         return exemplarsSingleView.profile ? (
-          <FlamegraphRenderer
-            showCredit={false}
-            profile={exemplarsSingleView.profile}
-            colorMode={colorMode}
-            ExportData={
-              <ExportData
-                flamebearer={exemplarsSingleView.profile}
-                exportPNG
-                exportJSON
-                exportPprof
-                exportHTML
-              />
-            }
-          />
+          <Box>
+            <FlamegraphRenderer
+              showCredit={false}
+              profile={exemplarsSingleView.profile}
+              colorMode={colorMode}
+              ExportData={
+                <ExportData
+                  flamebearer={exemplarsSingleView.profile}
+                  exportPNG
+                  exportJSON
+                  exportPprof
+                  exportHTML
+                />
+              }
+            />
+          </Box>
         ) : null;
       }
 
@@ -128,12 +133,31 @@ function ExemplarsSingleView() {
     <div>
       <PageTitle title={formatTitle('Tracing single', query)} />
       <div className="main-wrapper">
-        <Toolbar />
+        <Toolbar
+          onSelectedApp={(query) => {
+            dispatch(setQuery(query));
+          }}
+        />
         <Box>
           <p className={styles.heatmapTitle}>Heatmap</p>
           {heatmap}
         </Box>
-        {exemplarsSingleView.heatmap ? <Box>{flamegraphRenderer}</Box> : null}
+        {!exemplarsSingleView.profile && exemplarsSingleView.heatmap && (
+          <Box>
+            <div className={styles.heatmapSelectionGuide}>
+              <StatusMessage
+                type="info"
+                message="Select an area in the heatmap to get started"
+              />
+              <img
+                className={styles.gif}
+                src={heatmapSelectionGif}
+                alt="heatmap-selection-gif"
+              />
+            </div>
+          </Box>
+        )}
+        {exemplarsSingleView.heatmap ? flamegraphRenderer : null}
       </div>
     </div>
   );