+++ webapp/javascript/redux/reducers/continuous.ts
@@ -17,6 +17,8 @@ import * as tagsService from '@webapp/services/tags';
 import * as annotationsService from '@webapp/services/annotations';
 import { RequestAbortedError } from '@webapp/services/base';
 import { appendLabelToQuery } from '@webapp/util/query';
+import { createBiggestInterval } from '@webapp/util/timerange';
+import { formatAsOBject, toUnixTimestamp } from '@webapp/util/formatDate';
 import type { RootState } from '@webapp/redux/store';
 import { addNotification } from './notifications';
 import { createAsyncThunk } from '../async-thunk';
@@ -91,8 +93,6 @@ type DiffView =
   | { type: 'loaded'; profile: Profile }
   | { type: 'reloading'; profile: Profile };
 
-type DiffView2 = ComparisonView;
-
 type TimelineState =
   | { type: 'pristine'; timeline: Timeline }
   | { type: 'loading'; timeline: Timeline }
@@ -113,6 +113,8 @@ export type TagsState =
   | {
       type: 'loaded';
       tags: Record<string, TagsData>;
+      from: number;
+      until: number;
     }
   | { type: 'failed'; tags: Record<string, TagsData> };
 
@@ -135,7 +137,6 @@ interface ContinuousState {
 
   singleView: SingleView;
   diffView: DiffView;
-  diffView2: DiffView2;
   comparisonView: ComparisonView;
   tagExplorerView: TagExplorerView;
   newAnnotation: NewAnnotationState;
@@ -172,10 +173,6 @@ const initialState: ContinuousState = {
   singleView: { type: 'pristine' },
   diffView: { type: 'pristine' },
   comparisonView: {
-    left: { type: 'pristine' },
-    right: { type: 'pristine' },
-  },
-  diffView2: {
     left: { type: 'pristine' },
     right: { type: 'pristine' },
   },
@@ -535,7 +532,11 @@ export const fetchDiffView = createAsyncThunk<
   return Promise.reject(res.error);
 });
 
-export const fetchTags = createAsyncThunk(
+export const fetchTags = createAsyncThunk<
+  { appName: string; tags: string[]; from: number; until: number },
+  Query,
+  { state: { continuous: ContinuousState } }
+>(
   'continuous/fetchTags',
   async (query: Query, thunkAPI) => {
     const appName = queryToAppName(query);
@@ -547,12 +548,20 @@ export const fetchTags = createAsyncThunk(
       );
     }
 
-    const res = await tagsService.fetchTags(query);
+    const state = thunkAPI.getState().continuous;
+    const timerange = biggestTimeRangeInUnix(state);
+    const res = await tagsService.fetchTags(
+      query,
+      timerange.from,
+      timerange.until
+    );
 
     if (res.isOk) {
       return Promise.resolve({
         appName: appName.value,
         tags: res.value,
+        from: timerange.from,
+        until: timerange.until,
       });
     }
 
@@ -565,10 +574,63 @@ export const fetchTags = createAsyncThunk(
     );
 
     return Promise.reject(res.error);
+  },
+  {
+    // If we already loaded the tags for that application
+    // And we are trying to load tags for a smaller range
+    // Skip it, since we most likely already have that data
+    condition: (query, thunkAPI) => {
+      const appName = queryToAppName(query);
+      if (appName.isNothing) {
+        throw Error(
+          `Query '${appName}' is not a valid app, and it can't have any tags`
+        );
+      }
+
+      const state = thunkAPI.getState().continuous;
+      const timerange = biggestTimeRangeInUnix(state);
+
+      const s = state.tags[appName.value];
+
+      // Haven't loaded yet
+      if (!s) {
+        return true;
+      }
+
+      // Already loading that tag
+      if (s.type === 'loading') {
+        return false;
+      }
+
+      // Any other state that's not loaded
+      if (s.type !== 'loaded') {
+        return true;
+      }
+
+      const isInRange = (target: number) => {
+        return target >= s.from && target <= s.until;
+      };
+
+      const isSmallerThanLoaded =
+        isInRange(timerange.from) && isInRange(timerange.until);
+
+      return !isSmallerThanLoaded;
+    },
   }
 );
 
-export const fetchTagValues = createAsyncThunk(
+export const fetchTagValues = createAsyncThunk<
+  {
+    appName: string;
+    label: string;
+    values: string[];
+  },
+  {
+    query: Query;
+    label: string;
+  },
+  { state: { continuous: ContinuousState } }
+>(
   'continuous/fetchTagsValues',
   async (payload: { query: Query; label: string }, thunkAPI) => {
     const appName = queryToAppName(payload.query);
@@ -580,9 +642,20 @@ export const fetchTagValues = createAsyncThunk(
       );
     }
 
+    const state = thunkAPI.getState().continuous.tags[appName.value];
+    if (!state || state.type !== 'loaded') {
+      return Promise.reject(
+        new Error(
+          `Trying to load label-values for an unloaded label. This is likely due to a race condition.`
+        )
+      );
+    }
+
     const res = await tagsService.fetchLabelValues(
       payload.label,
-      payload.query
+      payload.query,
+      state.from,
+      state.until
     );
 
     if (res.isOk) {
@@ -602,6 +675,36 @@ export const fetchTagValues = createAsyncThunk(
     );
 
     return Promise.reject(res.error);
+  },
+  {
+    condition: ({ query, label }, thunkAPI) => {
+      const appName = queryToAppName(query);
+      if (appName.isNothing) {
+        throw Error(
+          `Query '${appName}' is not a valid app, and it can't have any tags`
+        );
+      }
+
+      const tagState = thunkAPI.getState().continuous.tags[appName.value];
+      if (!tagState || tagState.type !== 'loaded') {
+        return false;
+      }
+
+      const tagValueState = tagState.tags[label];
+
+      // Haven't loaded yet
+      if (!tagValueState) {
+        return true;
+      }
+
+      // Already loading that tag
+      // Or we already loaded it
+      if (tagValueState.type === 'loading' || tagValueState.type === 'loaded') {
+        return false;
+      }
+
+      return true;
+    },
   }
 );
 
@@ -949,6 +1052,7 @@ export const continuousSlice = createSlice({
 
     builder.addCase(fetchTags.fulfilled, (state, action) => {
       // convert each
+      // TODO(eh-am): don't delete tags if we already have them
       const tags = action.payload.tags.reduce((acc, tag) => {
         acc[tag] = { type: 'pristine' };
         return acc;
@@ -956,6 +1060,8 @@ export const continuousSlice = createSlice({
 
       state.tags[action.payload.appName] = {
         type: 'loaded',
+        from: action.payload.from,
+        until: action.payload.until,
         tags,
       };
     });
@@ -1094,6 +1200,25 @@ export const selectAnnotationsOrDefault = (state: RootState) => {
   return [];
 };
 
+export const selectRanges = (rootState: RootState) => {
+  const state = rootState.continuous;
+
+  return {
+    left: {
+      from: state.leftFrom,
+      until: state.leftUntil,
+    },
+    right: {
+      from: state.rightFrom,
+      until: state.rightUntil,
+    },
+    regular: {
+      from: state.from,
+      until: state.until,
+    },
+  };
+};
+
 function getNextStateFromPending(
   prevState: 'pristine' | 'loading' | 'reloading' | 'loaded'
 ) {
@@ -1102,4 +1227,15 @@ function getNextStateFromPending(
   }
 
   return 'reloading';
+}
+
+function biggestTimeRangeInUnix(state: ContinuousState) {
+  return createBiggestInterval({
+    from: [state.from, state.leftFrom, state.rightFrom]
+      .map(formatAsOBject)
+      .map(toUnixTimestamp),
+    until: [state.until, state.leftUntil, state.leftUntil]
+      .map(formatAsOBject)
+      .map(toUnixTimestamp),
+  });
 }