+++ packages/pyroscope-flamegraph/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.25.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.25.2...@pyroscope/flamegraph@0.25.3) (2022-11-08)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+## [0.25.2](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.25.1...@pyroscope/flamegraph@0.25.2) (2022-11-06)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
 ## [0.25.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.25.0...@pyroscope/flamegraph@0.25.1) (2022-11-02)
 
 **Note:** Version bump only for package @pyroscope/flamegraph