+++ webapp/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.56.5](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.4...@pyroscope/webapp@1.56.5) (2022-11-08)
+
+
+### Bug Fixes
+
+* **webapp:** pass from,until when calling /label{-values} ([#1677](https://github.com/pyroscope-io/pyroscope/issues/1677)) ([a82077d](https://github.com/pyroscope-io/pyroscope/commit/a82077dd23148d839b75b46fb11b6b71492eec99))
+
+
+
+
+
+## [1.56.4](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.3...@pyroscope/webapp@1.56.4) (2022-11-08)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.56.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.2...@pyroscope/webapp@1.56.3) (2022-11-07)
+
+
+### Bug Fixes
+
+* make oauth work when baseUrl is set ([#1673](https://github.com/pyroscope-io/pyroscope/issues/1673)) ([6cc1a2a](https://github.com/pyroscope-io/pyroscope/commit/6cc1a2a75387b0ca4a75ac57191a6a352a072a43))
+
+
+
+
+
+## [1.56.2](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.1...@pyroscope/webapp@1.56.2) (2022-11-06)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.56.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.0...@pyroscope/webapp@1.56.1) (2022-11-06)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+# [1.56.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.55.1...@pyroscope/webapp@1.56.0) (2022-11-04)
+
+
+### Features
+
+* show gif when heatmap has no selection ([#1658](https://github.com/pyroscope-io/pyroscope/issues/1658)) ([2a3243d](https://github.com/pyroscope-io/pyroscope/commit/2a3243de46d27de5d9cbbc9f50dd1089ba517a5b))
+
+
+
+
+
 ## [1.55.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.55.0...@pyroscope/webapp@1.55.1) (2022-11-02)
 
 **Note:** Version bump only for package @pyroscope/webapp